+++ test/test_formats_geotiff.py
@@ -44,18 +44,18 @@ def test_output_data():
     assert isinstance(output.profile(tile), dict)
     # write
     try:
-        tile.data = np.ones((1, ) + tile.shape)*128
-        output.write(tile)
+        data = np.ones((1, ) + tile.shape)*128
+        output.write(tile, data)
         # tiles_exist
         assert output.tiles_exist(tile)
         # read
-        data = output.read(tile).data
+        data = output.read(tile)
         assert isinstance(data, np.ndarray)
         assert not data[0].mask.any()
     finally:
         shutil.rmtree(temp_dir, ignore_errors=True)
     # read empty
-    data = output.read(tile).data
+    data = output.read(tile)
     assert isinstance(data, np.ndarray)
     assert data[0].mask.all()
     # empty